chorus(shopify): code reorganization
