+++ shopify/backend/jest.config.json
@@ -1,0 +1,14 @@
+{
+  "preset": "ts-jest",
+  "transform": {
+    "^.+\\.(t|j)sx?$": "ts-jest"
+  },
+  "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
+  "moduleFileExtensions": ["ts", "tsx", "js", "jsx", "json", "node"],
+  "reporters": [
+    "default",
+    ["jest-junit", {"outputDirectory": "reports", "outputName": "report.xml"}],
+    "github-actions"
+  ],
+  "setupFiles": ["<rootDir>/jest.setup.js"]
+}
