+++ shopify/backend/jest.setup.js
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const crypto = require("crypto").webcrypto;
+global.crypto = crypto;
