+++ shopify/backend/package.json
@@ -2,19 +2,47 @@
   "name": "backend",
   "version": "0.0.0",
   "private": true,
+  "files": [
+    "dist/**/*"
+  ],
+  "types": "dist/index.d.ts",
+  "main": "dist/index.js",
+  "type": "module",
   "scripts": {
+    "start": "wrangler dev --local",
+    "build": "tsc --project tsconfig.build.json",
+    "format": "prettier --write \"src/**/*.{ts,js}\"",
+    "lint": "eslint --ignore-path .eslintignore --ext .js,.ts .",
+    "test": "jest --config jest.config.json",
+    "prepare": "npm run build",
+    "prepublishOnly": "npm test && npm run lint",
+    "preversion": "npm run lint",
+    "postversion": "npm run format",
     "deploy": "wrangler publish",
-    "start": "wrangler dev"
+    "clean": "tsc --build --clean"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20230419.0",
     "@tsndr/cloudflare-worker-router": "^2.3.3",
     "typescript": "^5.0.4",
-    "wrangler": "^3.0.0"
+    "wrangler": "^3.0.0",
+    "@babel/core": "^7.21.4",
+    "@babel/preset-env": "^7.21.4",
+    "@types/jest": "^29.5.0",
+    "@typescript-eslint/eslint-plugin": "^5.57.0",
+    "@typescript-eslint/parser": "^5.57.0",
+    "babel-jest": "^29.5.0",
+    "eslint": "^8.37.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
+    "eslint-plugin-n": "^15.7.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "jest": "^29.5.0",
+    "jest-junit": "^15.0.0",
+    "ts-jest": "^29.0.5"
   },
-  "dependencies": {
-    "@shopify/shopify-api": "^7.3.0",
-    "crypto-js": "^4.1.1",
-    "crypto-ts": "^1.0.2"
-  }
+  "dependencies": {}
 }
