--- shopify/backend/src/env.ts
@@ -1,6 +1,0 @@
-export default interface Env {
-    SHOPIFY_CLIENT_ID: string,
-    SHOPIFY_SECRET_KEY: string,
-
-    TIKI: KVNamespace
-}
