+++ shopify/backend/src/routes/customer.ts
@@ -1,0 +1,15 @@
+import { RouterHandler } from "@tsndr/cloudflare-worker-router";
+
+export const dataRequest: RouterHandler<Env> = async ({ res, req, env }) => {
+    res.body = JSON.stringify([
+        req.body,
+        req.headers
+    ])
+}
+
+export const redact: RouterHandler<Env> = async ({ res, req, env }) => {
+    res.body = JSON.stringify([
+        req.body,
+        req.headers
+    ])
+}
