--- shopify/backend/src/shopify.app.toml
@@ -1,4 +1,0 @@
-name = "TIKI"
-scopes = "read_orders,write_discounts"
-extension_directories = ["extensions/*"]
-web_directories = ["backend"]
