+++ shopify/backend/src/shopify/interface/access_token_resp.ts
@@ -1,0 +1,3 @@
+export interface AccessTokenResponse {
+    access_token: string
+}
