+++ shopify/backend/src/shopify/interface/curr_app_installation_resp.ts
@@ -1,0 +1,7 @@
+export interface CurrentAppInstallationResp {
+    data: {
+        currentAppInstallation: {
+            id: string,
+        }
+    }
+}
