+++ shopify/backend/src/shopify/interface/index.ts
@@ -1,0 +1,2 @@
+export type * from './access_token_resp'
+export type * from './curr_app_installation_resp'
