+++ shopify/backend/src/tiki/interface/create_app_resp.ts
@@ -1,0 +1,7 @@
+export interface TikiCreateAppResp {
+    appId: String,
+    name: String,
+    orgId: String,
+    modified: String,
+    created: String
+}
