+++ shopify/backend/src/tiki/interface/create_keys_resp.ts
@@ -1,0 +1,7 @@
+export interface TikiCreateKeysResp {
+    id: String,
+    created: String,
+    isPublic: boolean,
+    secret: String,
+    public: boolean
+}
