+++ shopify/backend/src/tiki/interface/index.ts
@@ -1,0 +1,4 @@
+export type * from './create_app_resp'
+export type * from './create_keys_resp'
+export type * from './login_req'
+export type * from './login_resp'
