+++ shopify/backend/src/tiki/interface/login_req.ts
@@ -1,0 +1,7 @@
+export interface TikiLoginReq {
+    grant_type: String,
+    client_id: String,
+    subject_token: String,
+    subject_token_type: String,
+    scope: String,
+}
