+++ shopify/backend/src/tiki/interface/login_resp.ts
@@ -1,0 +1,8 @@
+
+export interface TikiLoginResp {
+    access_token: String,
+    refresh_token: String,
+    scope: String,
+    token_type: String,
+    expires_in: Number
+}
