+++ shopify/backend/src/worker-configuration.d.ts
@@ -1,0 +1,5 @@
+interface Env {
+    SHOPIFY_APP_THEME_EXTENSION_UUID: string,
+    SHOPIFY_CLIENT_ID: string,
+    SHOPIFY_SECRET_KEY: string
+}
