+++ shopify/extensions/theme-block/assets/tiki_anon.js
@@ -1,0 +1,4 @@
+
+class TikiAnon {
+   
+}
